% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas_playlists.R
\docType{data}
\name{christmas_playlists}
\alias{christmas_playlists}
\title{Spotify Christmas Playlists Dataframe}
\format{A data frame with 226 rows and 26 variables
The variables are as follows:
\describe{
   \item{playlist_name}{playlist name}
   \item{playlist_uri}{playlist uri}
   \item{playlist_tracks_url}{playlist tracks url}
   \item{playlist_num_tracks}{number of tracks in a given playlist}
   \item{snapshot_id}{snapshot id}
   \item{playlist_img}{playlist img}
   \item{track_name}{track name}
   \item{track_uri}{track uri}
   \item{artist_name}{artist name}
   \item{album_name}{album}
   \item{album_img}{album img}
   \item{track_popularity}{track popularity}
   \item{daceability}{danceability}
   \item{energy}{energy}
   \item{key}{key}
   \item{loudness}{loudness}
   \item{mode}{mode}
   \item{speechiness}{speechiness}
   \item{acousticness}{acoustiness}
   \item{instrumentalness}{instrumentalness}
   \item{liveness}{liveness}
   \item{valence}{valence}
   \item{tempo}{tempo}
   \item{duration_ms}{duration_ms}
   \item{time_signature}{time_signature}
   \item{key_mode}{key_mode}

}}
\source{
The data was sourced from the Spotify app.
}
\usage{
christmas_playlists
}
\description{
A dataframe containing playlists, track names, and an assortment of numeric and non-numeric variables.
}
\examples{
head(christmas_playlists)
}
\keyword{datasets}
