% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_col.R
\name{counter_col}
\alias{counter_col}
\title{Counter Column}
\usage{
counter_col(df)
}
\arguments{
\item{df}{a dataframe of a user's playlists.}
}
\value{
a dataframe with a new column called counter.
}
\description{
\code{counter_col} takes a dataframe and creates a new column such that tracks are numbered
one through n per playlist.
}
\details{
This function uses \code{dplyr}.
}
\examples{
data(christmas_playlists)
counter_col(christmas_playlists)
}
\author{
Belen Rodriguez <brodriguez@wesleyan.edu>

Kim Pham <spham@wesleyan.edu>
}
