% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_joyplot.R
\name{create_joyplot}
\alias{create_joyplot}
\title{Create Joyplot}
\usage{
create_joyplot(data = data, param, color = c("blue", "green"))
}
\arguments{
\item{data}{a dataframe of a user's playlists.}

\item{param}{a numeric variable in data}

\item{color}{a character vector of colors}
}
\value{
a joyplot containing playlists and the distributions of a numeric variable.
}
\description{
\code{create_joyplot} takes a dataframe, one argument, and an optional color vector
to plot a joyplot containing playlists and the distributions of a numeric variable.
}
\details{
This function uses \code{ggplot2},\code{rlang},\code{dplyr}, and \code{ggridges}.
}
\examples{
data(christmas_playlists)
create_joyplot(christmas_playlists, valence, c("red","green"))

}
\author{
Belen Rodriguez <brodriguez@wesleyan.edu>

Kim Pham <spham@wesleyan.edu>
}
