% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lr.R
\name{create_lr}
\alias{create_lr}
\title{Create Linear Regression Model}
\usage{
create_lr(data = data, param, param2, line = "white",
  points = "green", theme = "dark", multi = T)
}
\arguments{
\item{data}{a dataframe of a user's playlists.}

\item{param}{a numeric variable in data}

\item{line}{a string with a color name}

\item{points}{a string with a color name}

\item{theme}{a string that has valid values "light" or "dark"}

\item{multi}{a boolean}

\item{param}{another numeric variable in data}
}
\value{
linear regression model(s) for each user's playlists or combined playlists.
}
\description{
\code{create_lr} takes a dataframe, two arguments, a line color, a point color, a theme,
and a boolean indicating if the user wants to see a multiple linear regression models per playlist.
}
\details{
This function uses \code{ggplot2} and \code{rlang}.
}
\examples{
data(christmas_playlists)
create_lr(
data = christmas_playlists,
param = tempo,
param2 = valence,
points = "darkgreen",
line = "red",
theme = "light",
multi = T
)
}
\author{
Belen Rodriguez <brodriguez@wesleyan.edu>

Kim Pham <spham@wesleyan.edu>
}
