% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mean_df.R
\name{create_mean_df}
\alias{create_mean_df}
\title{Create a Dataframe or Table Containing Mean Variable Values}
\usage{
create_mean_df(data = data, param, asc = T, kable = F,
  bg = "green", text = "red")
}
\arguments{
\item{data}{a dataframe of a user's playlists.}

\item{param}{a numeric variable in data}

\item{asc}{a boolean to indicate the order of the result.}

\item{kable}{a boolean to indicate if user's wants to use kable on the result.}

\item{bg}{a string with a color name}

\item{text}{a string with a color name}
}
\value{
a dataframe or table containing mean variable values
}
\description{
\code{create_mean_df} takes a dataframe, one argument, an ascending order boolean value, a kable table boolean value,
a background color, and a text color.
}
\details{
This function uses \code{knitr},\code{dplyr},\code{kableExtra}, and \code{rlang}.
}
\examples{
data(christmas_playlists)
create_mean_df(christmas_playlists, valence, asc = F, kable=F)
}
\author{
Belen Rodriguez <brodriguez@wesleyan.edu>

Kim Pham <spham@wesleyan.edu>
}
