% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mean_hist.R
\name{create_mean_hist}
\alias{create_mean_hist}
\title{Create a Histogram Containing Mean Variable Values}
\usage{
create_mean_hist(data = data, param, color = "pink",
  border = "darkred")
}
\arguments{
\item{data}{a dataframe of a user's playlists.}

\item{param}{a numeric variable in data}

\item{color}{a string with a color name.}

\item{border}{a string with a color name.}
}
\value{
a histogram containing mean variable values
}
\description{
\code{create_mean_hist} takes a dataframe, one argument, a color, and a border color.
}
\details{
This function uses \code{ggplot2},\code{rlang}, and \code{dplyr}.
}
\examples{
data(christmas_playlists)
create_mean_hist(christmas_playlists, energy, color = "red", border = "darkgreen")
}
\author{
Belen Rodriguez <brodriguez@wesleyan.edu>

Kim Pham <spham@wesleyan.edu>
}
