% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wordcloud.R
\name{create_wordcloud}
\alias{create_wordcloud}
\title{Create a Wordcloud}
\usage{
create_wordcloud(data = data, playlist, stop_vector = NA,
  max_words = 30, del_file = T)
}
\arguments{
\item{data}{a dataframe of a user's playlists}

\item{playlist}{a character string}

\item{stop_vector}{a vector}

\item{max_words}{an integer}

\item{del_file}{a boolean}
}
\value{
a wordcloud generated from rquery
}
\description{
\code{create_wordcloud} produces a wordcloud with the lyrics of songs in a playlist that
are already stored within the geniusR package.
}
\details{
This function takes in a string of a playlist, a stop_vector that can remove certain common words
like "yeah" and "like" (as well as any other words the user find as irrelevent), and del_file which is
a boolean value that asserts whether the user wants to delete the output text file that is generated within
the function and that stores all the lyrics that are accumulated for the purposes of generating
the word cloud.
}
\examples{
data(christmas_playlists)
suppressWarnings(create_wordcloud(data = christmas_playlists, "Christmas Classics", stop_vector = c("yeah", "like"), del_file = T))
}
\author{
Belen Rodriguez <brodriguez@wesleyan.edu>

Kim Pham <kpham@wesleyan.edu>
}
