% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\docType{data}
\name{definitions}
\alias{definitions}
\title{Definitions Dataframe}
\format{A data frame with 13 rows and three variables
The variables are as follows:
\describe{
   \item{KEY}{name of parameter}
   \item{VALUE TYPE}{type int or float}3
   \item{VALUE DESCRIPTION}{description of parameter from Spotify API}

}}
\source{
The data was sourced from \href{(https://developer.spotify.com/documentation/web-api/reference/object-model/}{here}.
}
\usage{
definitons
}
\description{
A dataframe containing the Spotify parameters in the Spotify API, their value type (int or float),
and their description as it's shown in the Spotify API.
}
\examples{
head(definitions)
}
\keyword{datasets}
