% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_attributes.R
\name{feat_attributes}
\alias{feat_attributes}
\title{Get Feature Attributes}
\usage{
feat_attributes()
}
\value{
description in console
}
\description{
\code{feat_attributes} returns a description in the console with the contents of the definitions.Rda file.
}
\details{
This function does not take in any arguments. It is called on its own and returns
a description in the console with the contents of the definitions.Rda file.
}
\examples{
feat_attributes()
}
\author{
Belen Rodriguez <brodriguez@wesleyan.edu>

Kim Pham <kpham@wesleyan.edu>
}
