% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_names.R
\name{feat_names}
\alias{feat_names}
\title{Get Feature (Variable) Names}
\usage{
feat_names()
}
\value{
a character vector
}
\description{
\code{feat_names} produces a character vector with the names of all parameters.
}
\details{
This function does not take in any arguments, it just returns a character vector with the
names of all parameters when you type in feat_names().
}
\examples{
feat_names()
}
\author{
Belen Rodriguez <brodriguez@wesleyan.edu>

Kim Pham <kpham@wesleyan.edu>
}
