% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_ss.R
\name{filter_by_ss}
\alias{filter_by_ss}
\title{Filter Data by a Substring}
\usage{
filter_by_ss(data = data, col = playlist_name, substring)
}
\arguments{
\item{data}{a dataframe of a user's playlists}

\item{col}{playlist_name column set as default}

\item{substring}{a character string}
}
\value{
a data frame
}
\description{
\code{filter_by_ss} locates the desired substring in the column playlist_name from the Christmas playlist
dataframe.
}
\details{
This function takes in a character string and subsets all the playlist names in the playlist_name
column that contains that substring.
}
\examples{
data(christmas_playlists)
filter_by_ss(data = christmas_playlists, col=playlist_name, substring="Jazz")
}
\author{
Belen Rodriguez <brodriguez@wesleyan.edu>

Kim Pham <kpham@wesleyan.edu>
}
