% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist_top_songs.R
\name{playlist_top_songs}
\alias{playlist_top_songs}
\title{Get Top Songs From a User Playlist}
\usage{
playlist_top_songs(data = data, playlist, param, top = 5, asc = T,
  kable = T, bg = "azure", text = "palevioletred")
}
\arguments{
\item{data}{a dataframe of a user's playlists}

\item{playlist}{a character string}

\item{param}{a numeric variable in data}

\item{top}{an integer}

\item{asc}{a boolean}

\item{kable}{a boolean}

\item{bg}{a character string}

\item{text}{a character string}
}
\value{
a dataframe
}
\description{
\code{playlist_top_songs} produces a dataframe with the top 5 songs in a given playlist regarding the given
parameter.
}
\details{
This function takes in a string of a playlist and a string of parameter and returns a dataframe
with the top 5 songs
}
\examples{
data(christmas_playlists)
playlist_top_songs(data = christmas_playlists, "Christmas Jazz", valence, asc = T, kable = F, top = 10)
}
\author{
Belen Rodriguez <brodriguez@wesleyan.edu>

Kim Pham <kpham@wesleyan.edu>
}
